onvert some stuff from double to float.

+++ src/main/java/com/heroku/java/GettingStartedApplication.java
@@ -77,11 +77,11 @@ public class GettingStartedApplication {
                         if (Math.abs(h) + g == Math.abs(h)) {
                             t = a[ip][iq] / h;
                         } else {
-                            float theta = 0.5 * h / a[ip][iq];
-                            t = 1 / (Math.abs(theta) + Math.sqrt(1 + Math.pow(theta, 2)));
+                            float theta = (float) 0.5 * h / a[ip][iq];
+                            t = (float) 1 / (Math.abs(theta) + Math.sqrt(1 + Math.pow(theta, 2)));
                             if (theta < 0) t = -t;
                         }
-                        float c = 1 / Math.sqrt(1 + Math.pow(t, 2));
+                        float c = (float) 1 / Math.sqrt(1 + Math.pow(t, 2));
                         float s = t * c;
                         float tau = s / (1 + c);
                         h = t * a[ip][iq];
